BackEnd/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.6.0
dnspython>=2.6.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=43.0.0
email-validator>=2.2.0
python-multipart>=0.0.9
